mk-fnc.sh avoiding unwanted overwrite of existing components

+++ extras/mk-fnc.sh
@@ -9,15 +9,38 @@ function _help ()
     echo "USAGE: $( basename $0 ) -s|--sources <src_dir|file> <src_dir|file>..."
 }
 
+function _overwrite_file ()
+{
+	if [ -e "$1" ]; then
+		local T="$1"
+		echo -n "File $1 exists. Overwrite it ? [Y/N]: "
+		while read QT; do 
+			if [ "$QT" = "N" -o "$QT" = "n" ]; then
+		  		echo "$2 skipped..."
+				return 0
+			elif [ "$QT" = "Y" -o "$QT" = "y"  ]; then
+				return 1
+			else
+				echo -n "Not understood. [Y/N]: "
+			fi
+		done
+	else
+		return 1
+	fi
+}
+
 function _generate_with_template ()
 {
 	CZ="$1"
 	C=$(get_tool $CZ)
 	SK="$DEST/build_$C.sh"
+
+	_overwrite_file $SK $CZ
+	[ $? = 0 ] && return 1
+
 	echo "generating $SK"
 	cp $CD/build_script_template $SK
-	chmod +x $SK
-
+	chmod +x $SK	
 	sed -i s/"@_COMPONENT_"/$1/g $SK
 }
 
@@ -26,6 +49,10 @@ function _generate ()
 	CZ="$1"
 	C=$(get_tool $CZ)
 	SK="$DEST/build_$C.sh"
+	
+	_overwrite_file $SK $CZ
+	[ $? -eq 0 ] && return 1
+
 	echo "generating $SK"
 	touch $SK
 	chmod +x $SK
